myems-modbus-tcp/requirements.txt
mysql-connector
modbus_tk
schedule
python-decouple
